v0.3.0 update
Re-generated against Xero OpenAPI Spec v2.0.0. Version bumped to v0.3.0

Added enums for:
DELETED => Repeating Invoice
VOIDED => Receipt
EXCLUSIVE, INCLUSIVE, NOTAX => QuoteLineAmountTypes
NONE => Organisation
Various enums => ExpenseClaim
AUTHORISED & DELETED => Status for Batch Payments

Type changes from double to float:
UnitPrice in Purchase
Quantity & UnitAmount Quantity
QuantityOnHand in Itmes

+++ Xero.NetStandard.OAuth2/Model/LineItem.cs
@@ -51,14 +51,14 @@ namespace Xero.NetStandard.OAuth2.Model
         /// </summary>
         /// <value>LineItem Quantity</value>
         [DataMember(Name="Quantity", EmitDefaultValue=false)]
-        public double? Quantity { get; set; }
+        public float? Quantity { get; set; }
 
         /// <summary>
         /// LineItem Unit Amount
         /// </summary>
         /// <value>LineItem Unit Amount</value>
         [DataMember(Name="UnitAmount", EmitDefaultValue=false)]
-        public double? UnitAmount { get; set; }
+        public float? UnitAmount { get; set; }
 
         /// <summary>
         /// See Items
